GMAIL MANAGEMENT/requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
requests==2.32.3
openpyxl==3.1.5
